adds simple TomQueue activatation detection
